Split presentation and container components
