--- src/containers/Detail.js
@@ -1,7 +1,0 @@
-export default function Detail() {
-  return (
-    <div>
-      <h1>Detail</h1>
-    </div>
-  );
-}
